refactor(core): move i18n logic into its dedicated package
Move the i18n-related logic out of the instructions folder into
the i18n-dedicated folder / package.
