+++ packages/core/src/hydration/i18n.ts
@@ -8,7 +8,7 @@
 
 import {inject, Injector} from '../di';
 import {isRootTemplateMessage} from '../render3/i18n/i18n_util';
-import {createIcuIterator} from '../render3/instructions/i18n_icu_container_visitor';
+import {createIcuIterator} from '../render3/i18n/i18n_icu_container_visitor';
 import {I18nNode, I18nNodeKind, I18nPlaceholderType, TI18n, TIcu} from '../render3/interfaces/i18n';
 import {isTNodeShape, TNode, TNodeType} from '../render3/interfaces/node';
 import type {Renderer} from '../render3/interfaces/renderer';
