+++ packages/core/src/render3/i18n/i18n_icu_container_visitor.ts
@@ -9,7 +9,7 @@
 import {assertDomNode, assertNumber, assertNumberInRange} from '../../util/assert';
 import {EMPTY_ARRAY} from '../../util/empty';
 import {assertTIcu, assertTNodeForLView} from '../assert';
-import {getCurrentICUCaseIndex} from '../i18n/i18n_util';
+import {getCurrentICUCaseIndex} from './i18n_util';
 import {I18nRemoveOpCodes, TIcu} from '../interfaces/i18n';
 import {TIcuContainerNode} from '../interfaces/node';
 import {RNode} from '../interfaces/renderer_dom';
